Remove canvas state
